untitled folder/requirements.txt
aiohttp==3.9.1
uvloop==0.19.0
cchardet==2.1.7
aiodns==3.1.1
aiohttp-socks==0.8.4
cryptography==41.0.5
pyjwt==2.8.0
redis==5.0.1
prometheus-client==0.19.0
pyyaml==6.0.1
